Add Support for `*.net.cs` File Type
