+++ src/CommunityToolkit.Maui.Core/Handlers/DrawingView/DrawingViewHandler.shared.cs
@@ -180,79 +180,4 @@ public partial class DrawingViewHandler : ViewHandler<IDrawingView, MauiDrawingV
 		VirtualView.Lines.CollectionChanged += OnVirtualViewLinesCollectionChanged;
 	}
 }
-
-#else
-public partial class DrawingViewHandler : ViewHandler<IDrawingView, object>, IDrawingViewHandler
-{
-	/// <inheritdoc />
-	protected override object CreatePlatformView() => throw new NotSupportedException();
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.Lines"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapLines(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.ShouldClearOnFinish"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapShouldClearOnFinish(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.LineColor"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapLineColor(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.LineWidth"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapLineWidth(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.IsMultiLineModeEnabled"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapIsMultiLineModeEnabled(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView <see cref="IDrawingView.DrawAction"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapDrawAction(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the DrawingView Background property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="DrawingViewHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IDrawingView"/>.</param>
-	public static void MapDrawingViewBackground(DrawingViewHandler handler, IDrawingView view)
-	{
-	}
-
-	/// <inheritdoc/>
-	public void SetDrawingLineAdapter(IDrawingLineAdapter drawingLineAdapter)
-	{
-	}
-}
 #endif
