+++ src/CommunityToolkit.Maui.Core/Handlers/Popup/PopupHandler.shared.cs
@@ -48,77 +48,3 @@ public partial class PopupHandler
 	{
 	}
 }
-
-#if !(IOS || ANDROID || MACCATALYST || WINDOWS)
-public partial class PopupHandler : Microsoft.Maui.Handlers.ElementHandler<IPopup, object>
-{
-	/// <inheritdoc/>
-	protected override object CreatePlatformElement() => throw new NotSupportedException();
-
-	/// <summary>
-	/// Action that's triggered when the Popup is closed.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	/// <param name="result">The result that should return from this Popup.</param>
-	public static void MapOnClosed(PopupHandler handler, IPopup view, object? result)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup is Opened.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	/// <param name="result">We don't need to provide the result parameter here.</param>
-	public static void MapOnOpened(PopupHandler handler, IPopup view, object? result)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup is dismissed by tapping outside of the Popup.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	/// <param name="result">The result that should return from this Popup.</param>
-	public static void MapOnDismissedByTappingOutsideOfPopup(PopupHandler handler, IPopup view, object? result)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup <see cref="IPopup.Anchor"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	public static void MapAnchor(PopupHandler handler, IPopup view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup <see cref="IPopup.CanBeDismissedByTappingOutsideOfPopup"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	public static void MapCanBeDismissedByTappingOutsideOfPopup(PopupHandler handler, IPopup view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup <see cref="IPopup.Color"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	public static void MapColor(PopupHandler handler, IPopup view)
-	{
-	}
-
-	/// <summary>
-	/// Action that's triggered when the Popup <see cref="IPopup.Size"/> property changes.
-	/// </summary>
-	/// <param name="handler">An instance of <see cref="PopupHandler"/>.</param>
-	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
-	public static void MapSize(PopupHandler handler, IPopup view)
-	{
-	}
-}
-#endif
