+++ src/CommunityToolkit.Maui.Core/Views/DrawingView/PlatformView/MauiDrawingViewExtensions.shared.cs
@@ -63,21 +63,6 @@ public static class MauiDrawingViewExtensions
 		return smoothedPointsList.ToObservableCollection();
 	}
 
-	static Point GetIntermediatePoint(PointF p0, PointF p1, PointF p2, PointF p3, in float t, in float tt, in float ttt) => new()
-	{
-		X = 0.5f *
-			(2f * p1.X +
-				(p2.X - p0.X) * t +
-				(2f * p0.X - 5f * p1.X + 4f * p2.X - p3.X) * tt +
-				(3f * p1.X - p0.X - 3f * p2.X + p3.X) * ttt),
-		Y = 0.5f *
-			(2 * p1.Y +
-				(p2.Y - p0.Y) * t +
-				(2 * p0.Y - 5 * p1.Y + 4 * p2.Y - p3.Y) * tt +
-				(3 * p1.Y - p0.Y - 3 * p2.Y + p3.Y) * ttt)
-	};
-
-#if ANDROID || IOS || MACCATALYST || WINDOWS
 	/// <summary>
 	/// Set MultiLine mode
 	/// </summary>
@@ -187,5 +172,18 @@ public static class MauiDrawingViewExtensions
 			drawingView.Lines.Add(adapter.ConvertMauiDrawingLine(line));
 		}
 	}
-#endif
+
+	static Point GetIntermediatePoint(PointF p0, PointF p1, PointF p2, PointF p3, in float t, in float tt, in float ttt) => new()
+	{
+		X = 0.5f *
+		(2f * p1.X +
+			(p2.X - p0.X) * t +
+			(2f * p0.X - 5f * p1.X + 4f * p2.X - p3.X) * tt +
+			(3f * p1.X - p0.X - 3f * p2.X + p3.X) * ttt),
+		Y = 0.5f *
+		(2 * p1.Y +
+			(p2.Y - p0.Y) * t +
+			(2 * p0.Y - 5 * p1.Y + 4 * p2.Y - p3.Y) * tt +
+			(3 * p1.Y - p0.Y - 3 * p2.Y + p3.Y) * ttt)
+	};
 }
