+++ src/CommunityToolkit.Maui/Alerts/Toast/Toast.macios.cs
@@ -9,7 +9,22 @@ namespace CommunityToolkit.Maui.Alerts;
 
 public partial class Toast
 {
-	private static partial void DismissPlatform(CancellationToken token)
+	static CommunityToolkit.Maui.Core.Views.PlatformToast? PlatformToast { get; set; }
+
+	/// <summary>
+	/// Dispose Toast
+	/// </summary>
+	protected virtual void Dispose(bool isDisposing)
+	{
+		if (isDisposed)
+		{
+			return;
+		}
+
+		isDisposed = true;
+	}
+
+	static void DismissPlatform(CancellationToken token)
 	{
 		if (PlatformToast is not null)
 		{
@@ -22,7 +37,7 @@ public partial class Toast
 	/// <summary>
 	/// Show Toast
 	/// </summary>
-	private partial void ShowPlatform(CancellationToken token)
+	void ShowPlatform(CancellationToken token)
 	{
 		DismissPlatform(token);
 		token.ThrowIfCancellationRequested();
